backend/requirements.txt
alembic==1.12.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
arxiv==2.0.0
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==4.0.1
billiard==4.2.4
celery==5.3.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crossref-commons==0.0.7
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.131.0
feedparser==6.0.10
greenlet==3.3.2
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
openai==1.3.0
packaging==26.0
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.1.1
pip==26.0.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.1.0
PyPDF2==3.0.1
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.3
ratelimit==2.2.1
redis==5.0.1
requests==2.31.0
rsa==4.9.1
setuptools==65.5.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.52.1
tqdm==4.67.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0